realtime-portfolio-analysis/backend/requirements.txt
aiohttp
python-dotenv
loguru
Pillow
fastapi
pipecat-ai==0.0.67
rich
uvicorn
google-cloud-aiplatform
pymupdf
rich
colorama
pipecat-ai[google]
pipecat-ai[openai]
pipecat-ai[silero]
pipecat-ai[azure]
ipython
pandas
opencensus-ext-azure
azure-storage-blob
azure-search-documents
tavily-python
openpyxl
SQLAlchemy
yfinance